+++ packages/lumx-react/src/components/lightbox/Lightbox.stories.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable react-hooks/rules-of-hooks,react/display-name */
 import React from 'react';
-import { ImageBlock, Alignment, Lightbox, Button } from '@lumx/react';
+import { ImageBlock, Alignment, Lightbox, Button, Slideshow, SlideshowItem } from '@lumx/react';
 import { useBooleanState } from '@lumx/react/hooks/useBooleanState';
-import { LANDSCAPE_IMAGES } from '@lumx/react/stories/controls/image';
+import { LANDSCAPE_IMAGES, LANDSCAPE_IMAGES_ALT } from '@lumx/react/stories/controls/image';
 
 export default {
     title: 'LumX components/lightbox/Lightbox',
@@ -28,9 +28,17 @@ export default {
 /**
  * Base LightBox with image block
  */
-export const ImageBlock_ = {
+export const Image = {
     args: {
-        children: <ImageBlock align={Alignment.center} alt="" fillHeight image={LANDSCAPE_IMAGES.landscape1} />,
+        'aria-label': 'Fullscreen image',
+        children: (
+            <ImageBlock
+                align={Alignment.center}
+                fillHeight
+                image={LANDSCAPE_IMAGES.landscape1}
+                alt={LANDSCAPE_IMAGES_ALT.landscape1}
+            />
+        ),
     },
 };
 
@@ -39,7 +47,55 @@ export const ImageBlock_ = {
  */
 export const WithCloseButton = {
     args: {
-        ...ImageBlock_.args,
+        ...Image.args,
         closeButtonProps: { label: 'Close' },
+    },
+};
+
+/**
+ * Demo a LightBox containing an image slideshow
+ */
+export const ImageSlideshow = {
+    args: {
+        'aria-label': 'Fullscreen image slideshow',
+        closeButtonProps: { label: 'Close' },
+        children: (
+            <Slideshow
+                aria-label="Image slideshow"
+                theme="dark"
+                slideshowControlsProps={{
+                    nextButtonProps: { label: 'Next image' },
+                    previousButtonProps: { label: 'Previous image' },
+                }}
+                slideGroupLabel={(currentGroup, totalGroup) => `${currentGroup} of ${totalGroup}`}
+            >
+                <SlideshowItem>
+                    <ImageBlock
+                        align="center"
+                        fillHeight
+                        image={LANDSCAPE_IMAGES.landscape1}
+                        alt={LANDSCAPE_IMAGES_ALT.landscape1}
+                    />
+                </SlideshowItem>
+
+                <SlideshowItem>
+                    <ImageBlock
+                        align="center"
+                        fillHeight
+                        image={LANDSCAPE_IMAGES.landscape2}
+                        alt={LANDSCAPE_IMAGES_ALT.landscape2}
+                    />
+                </SlideshowItem>
+
+                <SlideshowItem>
+                    <ImageBlock
+                        align="center"
+                        fillHeight
+                        image={LANDSCAPE_IMAGES.landscape3}
+                        alt={LANDSCAPE_IMAGES_ALT.landscape3}
+                    />
+                </SlideshowItem>
+            </Slideshow>
+        ),
     },
 };